move distribution predicates to under common/tasks/sagemaker
